Revert "Revert "Rename lib to pyTransition""
