+++ README.md
@@ -1,25 +1,25 @@
 <!-- The fact that the library and the app have the same names is confusing.. -->
 # Transition
-Transition is a Python package designed to interact with the public API of the transit planning application Transition. It allows users to retrieve and request geographic and routing data from the app.\
+pyTransition is a Python package designed to interact with the public API of the transit planning application Transition. It allows users to retrieve and request geographic and routing data from the app.\
 The documentation for the Transition public API used by this library can be found [here](https://mathildebrosseau.github.io/transition-api/)
 
 ## Install and import Transition
-To install Transition, use the following command :
+To install pyTransition, use the following command :
 <!-- -[//]:#(probably something like pip install transition)- -->
 ```
-pip install transition-lib
+pip install pyTransition
 ```
-After installing Transition, it may be imported into Python code like :
+After installing pyTransition, it may be imported into Python code like :
 ```python
-import Transition
+from pyTransition.transition import Transition
 ```
 ## Usage
-Transition allows users to send HTTP requests. Before requesting data, users must first request a token. This is done using the `request_token` method. Afterwards, users can either set the token and server URL for the current instance using the `set_token` and `set_url` methods. The token and URL will be stored as local variables for the duration of the script only. Alternatively, users can send the token and the URL as parameters directly when calling methods.
+pyTransition allows users to send HTTP requests. Before requesting data, users must first request a token. This is done using the `request_token` method. Afterwards, users can either set the token and server URL for the current instance using the `set_token` and `set_url` methods. The token and URL will be stored as local variables for the duration of the script only. Alternatively, users can send the token and the URL as parameters directly when calling methods.
 
 All methods in Transition are static. The library provides the following :
 
 ### set_token :
-This method allows users to set the token that will be used for the API calls in a local variable within the Transition library.
+This method allows users to set the token that will be used for the API calls in a local variable within the pyTransition library.
 
 **Parameters :**&emsp;***token***&ensp;:&ensp;*string*\
 &emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;token used for API calls.
@@ -189,9 +189,9 @@ This method allows users to send calculation parameters to the Transition server
 &emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;If response code is not 200.
 
 ## Example
-Users can fetch the nodes which are currently loaded in the Transition application using Transition as follows :
+Users can fetch the nodes which are currently loaded in the Transition application using pyTransition as follows :
 ```python
-import Transition
+from pyTransition.transition import Transition
 
 def get_transition_nodes():
     url = "http://localhost:8080"
@@ -208,7 +208,7 @@ def get_transition_nodes():
 ```
 Alternatively, the token and URL can be set for the current session in order to avoid sending them as parameters. This can be useful if multiple calls are to be made in the script. This can be done as follows :
 ```python
-import Transition
+from pyTransition.transition import Transition
 
 def get_transition_nodes():
     # Set the URL and token.
@@ -223,9 +223,9 @@ def get_transition_nodes():
     print(nodes)
 ```
 
-Another example using Transition to get a new accessibility map :
+Another example using pyTransition to get a new accessibility map :
 ```python
-import Transition
+from pyTransition.transition import Transition
 from datetime import time
 import json
 
@@ -266,9 +266,9 @@ def get_transition_acessibility_map():
 
 ```
 
-Another example using Transition to get a new routes :
+Another example using pyTransition to get a new routes :
 ```python
-import Transition
+from pyTransition.transition import Transition
 from datetime import time
 import json
