+++ pyTransition/__init__.py
@@ -1,9 +1,9 @@
 """
-Transition.
+pyTransition
 
 A python package to interact with the Transition API.
 """
 
 __version__ = "0.1.0"
-__author__ = ''
-__credits__ = 'Polytechnique Montréal'
+__author__ = 'Transition City'
+__credits__ = 'Transition City'
