+++ pyTransition/transition.py
